github-commit-history-summarizer/pyproject.toml
[project]
name = "github-commit-summarizer"
version = "0.1.0"
description = "Generate AI-powered narrative summaries of GitHub repository commits"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.3.0",
]

[project.scripts]
commit-summarizer = "commit_summarizer.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/commit_summarizer"]
